报销Agent/requirements.txt
pandas
pillow
python-docx
pydantic
fastapi
uvicorn
python-multipart
opencv-python
json5
tabulate
numpy
matplotlib
requests
pytesseract
python-dotenv
streamlit>=1.30.0
watchdog
PyMuPDF>=1.18.0
pdf2image 